Duet wifi mounting bracket
Not sure what I'm going to do about cooling, but there's probably room
for thermal pads and an aluminum block under there.

Could probably drill through a chunk of aluminum and mount it on t-slots
with some thermal paste, then put some thermal pads on the bottom of it
to interface with the duet board near the drivers.