Allow Z motor wire connect to be more protected
Prior to this, the Z motor wire connector had to hang off of the printer
on the negative X side because of the front motor mount.

This way, the Z motor wire can be between the Z motor and the Y motor,
whcich is a bit better protected from little bumps.